Fix workaround for PIC servo bug
New behaviour:

On boot, we set the servos to 100% (this stops them pressing against the
stops and being damaged).

On both round end and code upload, we turn off the GPIOs and the motors.

On round end, we set the servos to 0. This isn't ideal -- we'd like to
leave them alone -- but since we can't disable power, it's the next-best
option.

On code upload, we set the servos to 100% (like what happens on boot).